Complete the serializer implementations.
